fix p command lol

+++ client/tecnicofs-client-api.c
@@ -14,19 +14,21 @@ int sockfd;
 socklen_t servlen;
 struct sockaddr_un serv_addr;
 
-int setSockAddrUn(char *path, struct sockaddr_un *addr) {
+int setSockAddrUn(char * path, struct sockaddr_un *addr) {
     if (addr == NULL)
         return 0;
 
     bzero((char *)addr, sizeof(struct sockaddr_un));
     addr->sun_family = AF_UNIX;
-    strcpy(addr->sun_path, path);
+    
+    if(path == NULL) mktemp(addr->sun_path);
+    else strcpy(addr->sun_path,path);
 
     return SUN_LEN(addr);
 }
 
 int tfsCreate(char *filename, char nodeType) {
-    char out_buffer[MAX_INPUT_SIZE+6];
+    char out_buffer[MAX_INPUT_SIZE];
     char in_buffer[MAX_INPUT_SIZE];
     
     sprintf(out_buffer,"c %s %c\n",filename,nodeType);
@@ -126,18 +128,13 @@ int tfsPrint(char *path) {
 int tfsMount(char * sockPath) {
     struct sockaddr_un client_addr;
     socklen_t clilen;
-    char path[MAX_PATH_LEN];
-
-    sprintf(path, "/tmp/%s%d.sock", SOCKET_NAME, getpid());
 
     if ((sockfd = socket(AF_UNIX, SOCK_DGRAM, 0)) < 0) {
         perror("server: can't open socket");
         return TECNICOFS_ERROR_OPEN_SESSION;
     }
 
-    unlink(path);
-
-    clilen = setSockAddrUn (path, &client_addr);
+    clilen = setSockAddrUn(NULL, &client_addr);
 
     if (bind(sockfd, (struct sockaddr *) &client_addr, clilen) < 0) {
         perror("server: bind error");
@@ -150,7 +147,9 @@ int tfsMount(char * sockPath) {
 }
 
 int tfsUnmount() {
-    close(sockfd);
+    if(close(sockfd) < 0) return FAIL;
+
     unlink(SOCKET_NAME);
-    return 0;
+    
+    return SUCCESS;
 }
